refactor(package): move ItemListView into view/utils
