+++ app/src/main/java/com/example/ezvault/utils/filter/ItemListFilter.java
@@ -3,7 +3,7 @@ package com.example.ezvault.utils.filter;
 import androidx.annotation.NonNull;
 
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.utils.ItemListView;
+import com.example.ezvault.view.utils.ItemListView;
 
 import java.util.ArrayList;
 import java.util.Iterator;
