+++ app/src/main/java/com/example/ezvault/view/adapter/ItemAdapter.java
@@ -18,7 +18,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.example.ezvault.R;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.utils.ItemListView;
+import com.example.ezvault.view.utils.ItemListView;
 
 import java.util.ArrayList;
 import java.util.List;
