+++ app/src/main/java/com/example/ezvault/view/utils/ItemListView.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.model.utils;
+package com.example.ezvault.view.utils;
 
 import com.example.ezvault.model.Item;
